\name{BaseXClient-package}
\alias{BaseXClient-package}
\alias{BaseXClient}
\docType{package}
\title{
Client library for the BaseX XML Database
}
\description{
BaseX is an XML database with a client-server architecture. It has support for XPath and XQuery 3.0,
as well as high performance indexing. This package implements the BaseX client library interface for R. 
}
\details{
\tabular{ll}{
Package: \tab BaseXClient\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2013-10-31\cr
License: \tab GPLv2\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
David Nusinow

Maintainer: David Nusinow <david@gravitypulls.net>
}
\references{
  \itemize{
    \item \href{http://docs.basex.org/wiki/Server_Protocol}{BaseX Server Protocol}
    \item \href{http://docs.basex.org/wiki/Standard_Mode}{BaseX Standard Mode Client Interaction}
    \item \href{http://docs.basex.org/wiki/Query_Mode}{BaseX Query Mode Client Interaction}
    \item \href{http://docs.basex.org/wiki/Clients}{Other BaseX Clients}
  }
}

\keyword{ package }
\keyword{ database }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
